+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.2.1] - 2026-08-28
+
 ### Fixed
 
 - `mpvd pick` won't submit when cancelling.
@@ -111,7 +113,8 @@ and this project adheres to
 - `remove`/`rm` subcommand
 - `send` subcommand
 
-[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.0...main
+[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.1...main
+[2.2.1]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.0.0...v2.1.0